+++ ckks/encoder.go
@@ -119,58 +119,56 @@ func NewEncoder(params *Parameters) Encoder {
 	}
 }
 
-// EncodeNew encodes a slice of complex128 of length slots = 2^{n} on a plaintext at the maximum level.
-func (encoder *encoderComplex128) EncodeNew(values []complex128, slots uint64) (plaintext *Plaintext) {
-	return encoder.EncodeAtLvlNew(encoder.params.MaxLevel(), values, slots)
+// EncodeNew encodes a slice of complex128 of length slots = 2^{logSlots} on new plaintext at the maximum level.
+func (encoder *encoderComplex128) EncodeNew(values []complex128, logSlots uint64) (plaintext *Plaintext) {
+	return encoder.EncodeAtLvlNew(encoder.params.MaxLevel(), values, logSlots)
 }
 
-// EncodeAtLvlNew encodes a slice of complex128 of length slots = 2^{n} on a plaintext at the desired level.
-func (encoder *encoderComplex128) EncodeAtLvlNew(level uint64, values []complex128, slots uint64) (plaintext *Plaintext) {
+// EncodeAtLvlNew encodes a slice of complex128 of length slots = 2^{logSlots} on new plaintext at the desired level.
+func (encoder *encoderComplex128) EncodeAtLvlNew(level uint64, values []complex128, logSlots uint64) (plaintext *Plaintext) {
 	plaintext = NewPlaintext(encoder.params, level, encoder.params.scale)
-	encoder.Encode(plaintext, values, slots)
+	encoder.Encode(plaintext, values, logSlots)
 	return
 }
 
-// EncodeNTTNew encodes a slice of complex128 of length slots = 2^{n} on a plaintext at the maximum level.
-// Returns a plaintext in the NTT domain.
-func (encoder *encoderComplex128) EncodeNTTNew(values []complex128, slots uint64) (plaintext *Plaintext) {
-	return encoder.EncodeNTTAtLvlNew(encoder.params.MaxLevel(), values, slots)
-}
-
-// EncodeNTTAtLvlNew encodes a slice of complex128 of length slots = 2^{n} on a plaintext at the desired level.
-// Returns a plaintext in the NTT domain.
-func (encoder *encoderComplex128) EncodeNTTAtLvlNew(level uint64, values []complex128, slots uint64) (plaintext *Plaintext) {
-	plaintext = NewPlaintext(encoder.params, encoder.params.MaxLevel(), encoder.params.scale)
-	encoder.EncodeNTT(plaintext, values, slots)
-	return
-}
-
-// Encode encodes a slice of complex128 of length slots = 2^{n} on the input plaintext.
-func (encoder *encoderComplex128) Encode(plaintext *Plaintext, values []complex128, slots uint64) {
-	encoder.Embed(values, slots)
+// Encode encodes a slice of complex128 of length slots = 2^{logSlots} on the input plaintext.
+func (encoder *encoderComplex128) Encode(plaintext *Plaintext, values []complex128, logSlots uint64) {
+	encoder.Embed(values, logSlots)
 	encoder.ScaleUp(plaintext.value, plaintext.scale, encoder.ringQ.Modulus[:plaintext.Level()+1])
 	encoder.WipeInternalMemory()
 	plaintext.isNTT = false
 }
 
-// EncodeNTT encodes a slice of complex128 of length slots = 2^{n} on the input plaintext.
+// EncodeNTTNew encodes a slice of complex128 of length slots = 2^{logSlots} on new plaintext at the maximum level.
 // Returns a plaintext in the NTT domain.
-func (encoder *encoderComplex128) EncodeNTT(plaintext *Plaintext, values []complex128, slots uint64) {
-	encoder.Encode(plaintext, values, slots)
+func (encoder *encoderComplex128) EncodeNTTNew(values []complex128, logSlots uint64) (plaintext *Plaintext) {
+	return encoder.EncodeNTTAtLvlNew(encoder.params.MaxLevel(), values, logSlots)
+}
+
+// EncodeNTTAtLvlNew encodes a slice of complex128 of length slots = 2^{logSlots} on new plaintext at the desired level.
+// Returns a plaintext in the NTT domain.
+func (encoder *encoderComplex128) EncodeNTTAtLvlNew(level uint64, values []complex128, logSlots uint64) (plaintext *Plaintext) {
+	plaintext = NewPlaintext(encoder.params, encoder.params.MaxLevel(), encoder.params.scale)
+	encoder.EncodeNTT(plaintext, values, logSlots)
+	return
+}
+
+// EncodeNTT encodes a slice of complex128 of length slots = 2^{logSlots} on the input plaintext.
+// Returns a plaintext in the NTT domain.
+func (encoder *encoderComplex128) EncodeNTT(plaintext *Plaintext, values []complex128, logSlots uint64) {
+	encoder.Encode(plaintext, values, logSlots)
 	encoder.ringQ.NTTLvl(plaintext.Level(), plaintext.value, plaintext.value)
 	plaintext.isNTT = true
 }
 
-// Embed encodes a vector and stores internaly the encoded values
+// Embed encodes a vector and stores internaly the encoded values.
 // To be used in conjonction with ScaleUp.
-func (encoder *encoderComplex128) Embed(values []complex128, slots uint64) {
+func (encoder *encoderComplex128) Embed(values []complex128, logSlots uint64) {
 
-	if uint64(len(values)) > encoder.params.N()/2 || uint64(len(values)) > slots {
-		panic("cannot Encode: too many values for the given number of slots")
-	}
+	slots := uint64(1 << logSlots)
 
-	if slots == 0 && slots&(slots-1) != 0 {
-		panic("cannot Encode: slots must be a power of two between 1 and N/2")
+	if uint64(len(values)) > encoder.params.N()/2 || uint64(len(values)) > slots || logSlots > encoder.params.LogN()-1 {
+		panic("cannot Encode: too many values/slots for the given ring degree")
 	}
 
 	for i := range values {
@@ -187,11 +185,12 @@ func (encoder *encoderComplex128) Embed(values []complex128, slots uint64) {
 	}
 }
 
-// ScaleUp writes the internaly stored encoded values on a polynomial
+// ScaleUp writes the internaly stored encoded values on a polynomial.
 func (encoder *encoderComplex128) ScaleUp(pol *ring.Poly, scale float64, moduli []uint64) {
 	scaleUpVecExact(encoder.valuesfloat, scale, moduli, pol.Coeffs)
 }
 
+// WipeInternalMemory sets the internaly stored encoded values of the encoder to zero.
 func (encoder *encoderComplex128) WipeInternalMemory() {
 	for i := range encoder.values {
 		encoder.values[i] = 0
@@ -213,7 +212,13 @@ func (encoder *encoderComplex128) Decode(plaintext *Plaintext, slots uint64) (re
 	return encoder.decodeAndRound(plaintext, slots, plaintext.scale)
 }
 
-func (encoder *encoderComplex128) decodeAndRound(plaintext *Plaintext, slots uint64, bound float64) (res []complex128) {
+func (encoder *encoderComplex128) decodeAndRound(plaintext *Plaintext, logSlots uint64, bound float64) (res []complex128) {
+
+	if logSlots > encoder.params.LogN()-1 {
+		panic("cannot Decode: too many slots for the given ring degree")
+	}
+
+	slots := uint64(1 << logSlots)
 
 	if plaintext.isNTT {
 		encoder.ringQ.InvNTTLvl(plaintext.Level(), plaintext.value, encoder.polypool)
@@ -506,52 +511,47 @@ func NewEncoderBigComplex(params *Parameters, logPrecision uint64) EncoderBigCom
 	}
 }
 
-// EncodeNew encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the maximum level.
-func (encoder *encoderBigComplex) EncodeNew(values []*ring.Complex, slots uint64) (plaintext *Plaintext) {
-	return encoder.EncodeAtLvlNew(encoder.params.MaxLevel(), values, slots)
+// EncodeNew encodes a slice of ring.Complex of length slots = 2^{logSlots} on a new plaintext at the maximum level.
+func (encoder *encoderBigComplex) EncodeNew(values []*ring.Complex, logSlots uint64) (plaintext *Plaintext) {
+	return encoder.EncodeAtLvlNew(encoder.params.MaxLevel(), values, logSlots)
 }
 
-// EncodeAtLvlNew encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the desired level.
-func (encoder *encoderBigComplex) EncodeAtLvlNew(level uint64, values []*ring.Complex, slots uint64) (plaintext *Plaintext) {
+// EncodeAtLvlNew encodes a slice of ring.Complex of length slots = 2^{logSlots} on a new plaintext at the desired level.
+func (encoder *encoderBigComplex) EncodeAtLvlNew(level uint64, values []*ring.Complex, logSlots uint64) (plaintext *Plaintext) {
 	plaintext = NewPlaintext(encoder.params, level, encoder.params.scale)
-	encoder.Encode(plaintext, values, slots)
+	encoder.Encode(plaintext, values, logSlots)
 	return
 }
 
-// EncodeNTTNew encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the maximum level.
+// EncodeNTTNew encodes a slice of ring.Complex of length slots = 2^{logSlots} on a plaintext at the maximum level.
 // Returns a plaintext in the NTT domain.
-func (encoder *encoderBigComplex) EncodeNTTNew(values []*ring.Complex, slots uint64) (plaintext *Plaintext) {
-	return encoder.EncodeNTTAtLvlNew(encoder.params.MaxLevel(), values, slots)
+func (encoder *encoderBigComplex) EncodeNTTNew(values []*ring.Complex, logSlots uint64) (plaintext *Plaintext) {
+	return encoder.EncodeNTTAtLvlNew(encoder.params.MaxLevel(), values, logSlots)
 }
 
-// EncodeNTTAtLvlNew encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the desired level.
+// EncodeNTTAtLvlNew encodes a slice of ring.Complex of length slots = 2^{logSlots} on a plaintext at the desired level.
 // Returns a plaintext in the NTT domain.
-func (encoder *encoderBigComplex) EncodeNTTAtLvlNew(level uint64, values []*ring.Complex, slots uint64) (plaintext *Plaintext) {
+func (encoder *encoderBigComplex) EncodeNTTAtLvlNew(level uint64, values []*ring.Complex, logSlots uint64) (plaintext *Plaintext) {
 	plaintext = NewPlaintext(encoder.params, encoder.params.MaxLevel(), encoder.params.scale)
-	encoder.EncodeNTT(plaintext, values, slots)
+	encoder.EncodeNTT(plaintext, values, logSlots)
 	return
 }
 
-// Encode encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the input plaintext level.
+// Encode encodes a slice of ring.Complex of length slots = 2^{logSlots} on a plaintext at the input plaintext level.
 // Returns a plaintext in the NTT domain.
-func (encoder *encoderBigComplex) EncodeNTT(plaintext *Plaintext, values []*ring.Complex, slots uint64) {
-
-	encoder.Encode(plaintext, values, slots)
-
+func (encoder *encoderBigComplex) EncodeNTT(plaintext *Plaintext, values []*ring.Complex, logSlots uint64) {
+	encoder.Encode(plaintext, values, logSlots)
 	encoder.ringQ.NTTLvl(plaintext.Level(), plaintext.value, plaintext.value)
-
 	plaintext.isNTT = true
 }
 
-// Encode encodes a slice of ring.Complex of length slots = 2^{n} on a plaintext at the input plaintext level.
-func (encoder *encoderBigComplex) Encode(plaintext *Plaintext, values []*ring.Complex, slots uint64) {
+// Encode encodes a slice of ring.Complex of length slots = 2^{logSlots} on a plaintext at the input plaintext level.
+func (encoder *encoderBigComplex) Encode(plaintext *Plaintext, values []*ring.Complex, logSlots uint64) {
 
-	if uint64(len(values)) > encoder.ringQ.N>>1 || uint64(len(values)) > slots {
-		panic("cannot Encode: too many values for the given number of slots")
-	}
+	slots := uint64(1 << logSlots)
 
-	if slots == 0 && slots&(slots-1) == 0 {
-		panic("cannot Encode: slots must be a power of two between 1 and N/2")
+	if uint64(len(values)) > encoder.params.N()/2 || uint64(len(values)) > slots || logSlots > encoder.params.LogN()-1 {
+		panic("cannot Encode: too many values/slots for the given ring degree")
 	}
 
 	if uint64(len(values)) != slots {
@@ -589,16 +589,22 @@ func (encoder *encoderBigComplex) Encode(plaintext *Plaintext, values []*ring.Co
 
 // Decode decodes the Plaintext values to a slice of complex128 values of size at most N/2.
 // Rounds the decimal part of the output (the bits under the scale) to "logPrecision" bits of precision.
-func (encoder *encoderBigComplex) DecodeAndRound(plaintext *Plaintext, slots uint64, bound float64) (res []*ring.Complex) {
-	return encoder.decodeAndRound(plaintext, slots, bound)
+func (encoder *encoderBigComplex) DecodeAndRound(plaintext *Plaintext, logSlots uint64, bound float64) (res []*ring.Complex) {
+	return encoder.decodeAndRound(plaintext, logSlots, bound)
 }
 
-func (encoder *encoderBigComplex) Decode(plaintext *Plaintext, slots uint64) (res []*ring.Complex) {
-	return encoder.decodeAndRound(plaintext, slots, plaintext.scale)
+func (encoder *encoderBigComplex) Decode(plaintext *Plaintext, logSlots uint64) (res []*ring.Complex) {
+	return encoder.decodeAndRound(plaintext, logSlots, plaintext.scale)
 }
 
 // Decode decodes the Plaintext values to a slice of complex128 values of size at most N/2.
-func (encoder *encoderBigComplex) decodeAndRound(plaintext *Plaintext, slots uint64, bound float64) (res []*ring.Complex) {
+func (encoder *encoderBigComplex) decodeAndRound(plaintext *Plaintext, logSlots uint64, bound float64) (res []*ring.Complex) {
+
+	slots := uint64(1 << logSlots)
+
+	if logSlots > encoder.params.LogN()-1 {
+		panic("cannot Decode: too many slots for the given ring degree")
+	}
 
 	encoder.ringQ.InvNTTLvl(plaintext.Level(), plaintext.value, encoder.polypool)
 	encoder.ringQ.PolyToBigint(encoder.polypool, encoder.bigintCoeffs)