+++ examples/ckks/sigmoid/main.go
@@ -67,8 +67,7 @@ func chebyshevinterpolation() {
 	fmt.Println()
 
 	// Plaintext creation and encoding process
-	plaintext := ckks.NewPlaintext(params, params.MaxLevel(), params.Scale())
-	encoder.Encode(plaintext, values, params.Slots())
+	plaintext := encoder.EncodeNew(values, params.LogSlots())
 
 	// Encryption process
 	var ciphertext *ckks.Ciphertext
@@ -79,6 +78,14 @@ func chebyshevinterpolation() {
 	// Evaluation process
 	// We approximate f(x) in the range [-8, 8] with a Chebyshev interpolant of 33 coefficients (degree 32).
 	chebyapproximation := ckks.Approximate(f, -8, 8, 33)
+
+	a := chebyapproximation.A()
+	b := chebyapproximation.B()
+
+	// Change of variable
+	evaluator.MultByConst(ciphertext, 2/(b-a), ciphertext)
+	evaluator.AddConst(ciphertext, (-a-b)/(b-a), ciphertext)
+	evaluator.Rescale(ciphertext, params.Scale(), ciphertext)
 
 	// We evaluate the interpolated Chebyshev interpolant on the ciphertext
 	if ciphertext, err = evaluator.EvaluateCheby(ciphertext, chebyapproximation, rlk); err != nil {
@@ -111,9 +118,7 @@ func round(x complex128) complex128 {
 
 func printDebug(params *ckks.Parameters, ciphertext *ckks.Ciphertext, valuesWant []complex128, decryptor ckks.Decryptor, encoder ckks.Encoder) (valuesTest []complex128) {
 
-	slots := uint64(len(valuesWant))
-
-	valuesTest = encoder.Decode(decryptor.DecryptNew(ciphertext), slots)
+	valuesTest = encoder.Decode(decryptor.DecryptNew(ciphertext), params.LogSlots())
 
 	fmt.Println()
 	fmt.Printf("Level: %d (logQ = %d)\n", ciphertext.Level(), params.LogQLvl(ciphertext.Level()))